Promises in place.
Did limited testing to see if got any errors but did not.
Errors on WHOIS lookup do get logged to console right now.
Using async-await so ecmaVersion bumped to 8.
Switched out 'node-whois' for 'whois-promise' since whois lookup needed to be promised.
WHOIS parsing moved to manager, no need for whois lookup anymore.
