+++ .eslintrc.yml
@@ -3,8 +3,8 @@ env:
   es6: true
   node: true
 extends:
-  - 'eslint:recommended'
-  - 'google'
-  - 'eslint-config-prettier'
+  - "eslint:recommended"
+  - "google"
+  - "eslint-config-prettier"
 parserOptions:
-  ecmaVersion: 2015
+  ecmaVersion: 8
