+++ lib/server.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const socketio = require("socket.io");
 const morgan = require("morgan");
+const fs = require("fs").promises;
 
 const app = express();
 const http = require("http").createServer(app);
@@ -26,16 +27,20 @@ io.on("connection", sockets.handler);
  */
 function startServer() {
   // Grab most current config settings
-  yamler.getObjectFromYaml(yamler.CONFIG_YAML_PATH, (err, config) => {
-    if (err) throw err;
-    const RUNNING_PORT = config.app.port;
-    // Run Web Server
-    http.listen(RUNNING_PORT, () => {
-      console.log(
-        `Web interface available at http://localhost:${RUNNING_PORT}/`
-      );
+  fs.readFile(yamler.CONFIG_YAML_PATH)
+    .then(yamler.parseObjectFromYaml)
+    .then((config) => {
+      const RUNNING_PORT = config.app.port;
+      // Run Web Server
+      http.listen(RUNNING_PORT, () => {
+        console.log(
+          `Web interface available at http://localhost:${RUNNING_PORT}/`
+        );
+      });
+    })
+    .catch((error) => {
+      throw error;
     });
-  });
 }
 
 /**
