+++ lib/sockets.js
@@ -1,3 +1,5 @@
+const fs = require("fs").promises;
+
 const yamler = require("./yamler");
 const constants = require("./constants");
 
@@ -20,17 +22,25 @@ function socketHandler(socket) {
   // Log connection
   logSocketTraffic(socket, "connection");
   // Send configuration to client on connection
-  yamler.getObjectFromYaml(yamler.CONFIG_YAML_PATH, (err, config) => {
-    if (err) throw err;
-    socket.emit(constants.SCK_LOADED_CFG, config);
-    logSocketTraffic(socket, "sent config");
-  });
+  fs.readFile(yamler.CONFIG_YAML_PATH)
+    .then(yamler.parseObjectFromYaml)
+    .then((configInfo) => {
+      socket.emit(constants.SCK_LOADED_CFG, configInfo);
+      logSocketTraffic(socket, "sent config");
+    })
+    .catch((reason) => {
+      throw reason;
+    });
   // Send domain list to client on connection
-  yamler.getObjectFromYaml(yamler.DOMAIN_YAML_PATH, (err, domains) => {
-    if (err) throw err;
-    socket.emit(constants.SCK_LOADED_DOMAINS, domains);
-    logSocketTraffic(socket, "sent domains");
-  });
+  fs.readFile(yamler.DOMAIN_YAML_PATH)
+    .then(yamler.parseObjectFromYaml)
+    .then((domainInfo) => {
+      socket.emit(constants.SCK_LOADED_DOMAINS, domainInfo);
+      logSocketTraffic(socket, "sent domains");
+    })
+    .catch((reason) => {
+      throw reason;
+    });
 
   // Hand events
   socket.on("disconnect", (reason) => {
@@ -38,55 +48,61 @@ function socketHandler(socket) {
   });
   socket.on(constants.SCK_UPDATE_DOMAIN, (val) => {
     logSocketTraffic(socket, `SCK_UPDATE_DOMAIN: ${JSON.stringify(val)}`);
-    yamler.getObjectFromYaml(yamler.DOMAIN_YAML_PATH, (err, domainInfo) => {
-      if (err) throw err;
-      const domainList = domainInfo.domains;
-      let updatedEntry = false;
-      for (let i = 0; i < domainList.length; i++) {
-        if (domainList[i].fqdn === val.fqdn) {
-          console.log(`SOCK: Matched updated domain to ${domainList[i].fqdn}`);
-          domainList[i].name = val.name;
-          domainList[i].alerts = val.alerts;
-          domainList[i].enabled = val.enabled;
-          updatedEntry = true;
+    fs.readFile(yamler.DOMAIN_YAML_PATH)
+      .then(yamler.parseObjectFromYaml)
+      .then((domainInfo) => {
+        const domainList = domainInfo.domains;
+        let updatedEntry = false;
+        for (let i = 0; i < domainList.length; i++) {
+          if (domainList[i].fqdn === val.fqdn) {
+            console.info(
+              `SOCK: Matched updated domain to ${domainList[i].fqdn}`
+            );
+            domainList[i].name = val.name;
+            domainList[i].alerts = val.alerts;
+            domainList[i].enabled = val.enabled;
+            updatedEntry = true;
+          }
         }
-      }
-      if (!updatedEntry) {
-        domainList.push(val);
-      }
-      yamler.setYamlFromObject(
-        yamler.DOMAIN_YAML_PATH,
-        { domains: domainList },
-        (err) => {
-          if (err) throw err;
-          socket.emit(constants.SCK_LOADED_DOMAINS, { domains: domainList });
-          logSocketTraffic(socket, "sent domain update");
+        if (!updatedEntry) {
+          domainList.push(val);
         }
-      );
-    });
+        const domainYaml = yamler.parseYamlFromObj({ domains: domainList });
+        return fs.writeFile(yamler.DOMAIN_YAML_PATH, domainYaml);
+      })
+      .then(fs.readFile(yamler.DOMAIN_YAML_PATH))
+      .then(yamler.getObjectFromYaml)
+      .then((domainInfo) => {
+        socket.emit(constants.SCK_LOADED_DOMAINS, domainInfo);
+        logSocketTraffic(socket, "sent domain update.");
+      })
+      .catch((error) => {
+        throw error;
+      });
   });
   socket.on(constants.SCK_DELETE_DOMAIN, (val) => {
     logSocketTraffic(socket, `SCK_DELETE_DOMAIN: ${JSON.stringify(val)}`);
-
-    yamler.getObjectFromYaml(yamler.DOMAIN_YAML_PATH, (err, domainInfo) => {
-      if (err) throw err;
-      const domainList = [];
-      for (let i = 0; i < domainInfo.domains.length; i++) {
-        if (domainInfo.domains[i].fqdn !== val.fqdn) {
-          domainList.push(domainInfo.domains[i]);
+    fs.readFile(yamler.DOMAIN_YAML_PATH)
+      .then(yamler.parseObjectFromYaml)
+      .then((domainInfo) => {
+        const domainList = [];
+        for (let i = 0; i < domainInfo.domains.length; i++) {
+          if (domainInfo.domains[i].fqdn !== val.fqdn) {
+            domainList.push(domainInfo.domains[i]);
+          }
         }
-      }
-
-      yamler.setYamlFromObject(
-        yamler.DOMAIN_YAML_PATH,
-        { domains: domainList },
-        (err) => {
-          if (err) throw err;
-          socket.emit(constants.SCK_LOADED_DOMAINS, { domains: domainList });
-          logSocketTraffic(socket, "sent domain update");
-        }
-      );
-    });
+        const domainYaml = yamler.parseYamlFromObj({ domains: domainList });
+        return fs.writeFile(yamler.DOMAIN_YAML_PATH, domainYaml);
+      })
+      .then(fs.readFile(yamler.DOMAIN_YAML_PATH))
+      .then(yamler.parseObjectFromYaml)
+      .then((domainInfo) => {
+        socket.emit(constants.SCK_LOADED_DOMAINS, domainInfo);
+        logSocketTraffic(socket, "sent domain update");
+      })
+      .catch((error) => {
+        throw error;
+      });
   });
 }
