+++ lib/whois-manager.js
@@ -1,8 +1,9 @@
 const path = require("path");
-const fs = require("fs");
+const fs = require("fs").promises;
+const whois = require("whois-promise");
+const existsSync = require("fs").existsSync;
 
 const yamler = require("./yamler");
-const whoislu = require("./whois-lookup");
 
 const NINE_MONTHS_IN_MILLIS = 9 * 30 * 24 * 60 * 60 * 1000;
 const THREE_MONTH_IN_MILLIS = 3 * 30 * 24 * 60 * 60 * 1000;
@@ -11,6 +12,8 @@ const ONE_MONTH_IN_MILLIS = 1 * 30 * 24 * 60 * 60 * 1000;
 const TWO_WEEKS_IN_MILLIS = 14 * 24 * 60 * 60 * 1000;
 const UPDATE_INTERVAL = 24 * 60 * 60 * 1000;
 const STARTUP_TIMEOUT = 1 * 60 * 1000;
+
+const ATTRIBUTE_REGEX = /(^[^:]+): (.*)/;
 
 // Start whois look ups after STARTUP TIMEOUT and rerun the whois update
 // every UPDATE_INTERVAL
@@ -23,78 +26,107 @@ setTimeout(function go() {
  * Run the whois updates
  */
 function doWhoisUpdates() {
-  yamler.getObjectFromYaml(yamler.DOMAIN_YAML_PATH, (err, obj) => {
-    if (err) throw err;
-    const domains = obj.domains;
-    for (let i = 0; i < domains.length; i++) {
-      if (domains[i].enabled) updateWhoisIfNeeded(domains[i].fqdn);
-    }
-  });
+  fs.readFile(yamler.DOMAIN_YAML_PATH)
+    .then(yamler.parseObjectFromYaml)
+    .then((domainInfo) => {
+      const whoisPromises = [];
+      const domains = domainInfo.domains;
+      for (let i = 0; i < domains.length; i++) {
+        if (domains[i].enabled)
+          whoisPromises.push(updateWhoisIfNeeded(domains[i].fqdn));
+      }
+      return Promise.all(whoisPromises);
+    })
+    .then((result) => {
+      console.debug("WHOIS: Cached entries: ", result);
+    })
+    .catch((error) => {
+      throw error;
+    });
 }
 
 /**
  *
  * @param {string} domain
- * @return {*}
+ * @return {Promise}
  */
 function updateWhoisIfNeeded(domain) {
   const ypath = yamlFilepathFromDomain(domain);
-  if (!fs.existsSync(ypath)) {
-    console.info(`WHOIS: No existing whois.yaml for ${domain}. Fetching..`);
-    return updateWhoisYamlFor(ypath, domain);
-  }
-  yamler.getObjectFromYaml(ypath, (err, obj) => {
-    if (err) throw err;
-    console.info(`WHOIS: Checking gather date for ${domain}`);
-    const updateTime = new Date(obj.whois_db_update_time);
-    let timeDelta = Math.abs(new Date() - updateTime);
-    console.info(
-      `WHOIS: Age of existing whois.yaml is ${humanReadableTime(
-        timeDelta
-      )} (${timeDelta})`
-    );
-    if (timeDelta >= NINE_MONTHS_IN_MILLIS) {
-      console.info("WHOIS: Age of whois.yaml > 9 months. Updating..");
-      return updateWhoisYamlFor(ypath, domain);
+  const updateWhoisPromise = function (resolve, reject) {
+    console.log(`creating whois promise for ${domain}`);
+    if (!existsSync(ypath)) {
+      console.info(`WHOIS: No existing whois.yaml for ${domain}. Fetching..`);
+      return resolve(updateWhoisYamlFor(ypath, domain));
     }
-    const expireTime = new Date(obj.registrar.registration_expiration);
-    timeDelta = Math.abs(new Date() - expireTime);
-    console.info(
-      `WHOIS: ${domain} expires in ${humanReadableTime(
-        timeDelta
-      )} on ${expireTime.toDateString()}`
-    );
-    const fudgedDelta = timeDelta - UPDATE_INTERVAL;
-    if (fudgedDelta < TWO_WEEKS_IN_MILLIS && timeDelta >= TWO_WEEKS_IN_MILLIS) {
-      console.info(
-        `WHOIS: Less than two weeks until expiriry. Updating whois.yaml`
-      );
-      return updateWhoisYamlFor(ypath, domain);
-    }
-    if (fudgedDelta < ONE_MONTH_IN_MILLIS && timeDelta >= ONE_MONTH_IN_MILLIS) {
-      console.info(
-        `WHOIS: Less than one month until expiriry. Updating whois.yaml`
-      );
-      return updateWhoisYamlFor(ypath, domain);
-    }
-    if (fudgedDelta < TWO_MONTH_IN_MILLIS && timeDelta >= TWO_MONTH_IN_MILLIS) {
-      console.info(
-        `WHOIS: Less than two months until expiriry. Updating whois.yaml`
-      );
-      return updateWhoisYamlFor(ypath, domain);
-    }
-    if (
-      fudgedDelta < THREE_MONTH_IN_MILLIS &&
-      timeDelta >= THREE_MONTH_IN_MILLIS
-    ) {
-      console.info(
-        `WHOIS: Less than three months until expiriry. Updating whois.yaml`
-      );
-      return updateWhoisYamlFor(ypath, domain);
-    }
-    console.info(`WHOIS: Within age boundary checks for ${domain} whois.yaml`);
-    return true;
-  });
+    console.info(`WHOIS: Validating cached whois data for ${domain}`);
+    fs.readFile(ypath)
+      .then(yamler.parseObjectFromYaml)
+      .then((whoisInfo) => {
+        const updateTime = new Date(whoisInfo.whois_db_update_time);
+        let timeDelta = Math.abs(new Date() - updateTime);
+        console.info(
+          `WHOIS: Age of existing whois.yaml is ${humanReadableTime(
+            timeDelta
+          )} (${timeDelta})`
+        );
+        if (timeDelta >= NINE_MONTHS_IN_MILLIS) {
+          console.info("WHOIS: Age of whois.yaml > 9 months. Updating..");
+          return resolve(updateWhoisYamlFor(ypath, domain));
+        }
+        const expireTime = new Date(
+          whoisInfo.registrar.registration_expiration
+        );
+        timeDelta = Math.abs(new Date() - expireTime);
+        console.info(
+          `WHOIS: ${domain} expires in ${humanReadableTime(
+            timeDelta
+          )} on ${expireTime.toDateString()}`
+        );
+        const fudgedDelta = timeDelta - UPDATE_INTERVAL;
+        if (
+          fudgedDelta < TWO_WEEKS_IN_MILLIS &&
+          timeDelta >= TWO_WEEKS_IN_MILLIS
+        ) {
+          console.info(
+            `WHOIS: Less than two weeks until expiriry. Updating whois.yaml`
+          );
+          return resolve(updateWhoisYamlFor(ypath, domain));
+        }
+        if (
+          fudgedDelta < ONE_MONTH_IN_MILLIS &&
+          timeDelta >= ONE_MONTH_IN_MILLIS
+        ) {
+          console.info(
+            `WHOIS: Less than one month until expiriry. Updating whois.yaml`
+          );
+          return resolve(updateWhoisYamlFor(ypath, domain));
+        }
+        if (
+          fudgedDelta < TWO_MONTH_IN_MILLIS &&
+          timeDelta >= TWO_MONTH_IN_MILLIS
+        ) {
+          console.info(
+            `WHOIS: Less than two months until expiriry. Updating whois.yaml`
+          );
+          return resolve(updateWhoisYamlFor(ypath, domain));
+        }
+        if (
+          fudgedDelta < THREE_MONTH_IN_MILLIS &&
+          timeDelta >= THREE_MONTH_IN_MILLIS
+        ) {
+          console.info(
+            `WHOIS: Less than three months until expiriry. Updating whois.yaml`
+          );
+          return resolve(updateWhoisYamlFor(ypath, domain));
+        }
+        console.info(
+          `WHOIS: Within age boundary checks for ${domain} whois.yaml`
+        );
+        return resolve(true);
+      });
+  };
+
+  return new Promise(updateWhoisPromise);
 }
 
 /**
@@ -140,22 +172,249 @@ function humanReadableTime(time) {
  *
  * @param {string} ypath yaml path
  * @param {string} domain
+ * @return {Promise}
  */
 function updateWhoisYamlFor(ypath, domain) {
-  whoislu.getWhoisFor(domain, (err, data) => {
-    if (err) {
-      console.error(`WHOIS: Failure looking up ${domain}`);
-      return console.error("WHOIS:", err);
-    }
-    if (!data.registrar || !data.registrar.registration_expiration) {
-      return console.error(
-        `WHOIS: Lookup returned empty data for ${domain}.`,
-        JSON.stringify(data)
-      );
-    }
-    yamler.setYamlFromObject(ypath, data, (err) => {
-      if (err) return console.error("WHOIS: YAML WRITE", err);
-      console.info(`WHOIS: Saved new whois.yaml for ${domain}`);
+  return whois
+    .raw(domain, {})
+    .then((data) => {
+      const whoisData = simplifyWhois(data);
+      return writeWhoisData(ypath, whoisData);
+    })
+    .catch((error) => {
+      console.error(error);
     });
+}
+
+/**
+ *
+ * @param {string} ypath
+ * @param {Object} whoisObj
+ * @return {Promise}
+ */
+function writeWhoisData(ypath, whoisObj) {
+  return new Promise((resolve, reject) => {
+    if (whoisObj === {} || Object.keys(whoisObj).length == 0) {
+      reject(new Error("WHOIS: Empty WHOIS Response"));
+    } else if (whoisObj.raw) {
+      reject(new Error(whoisObj.raw));
+    } else {
+      const whoisYaml = yamler.parseYamlFromObj(whoisObj);
+      resolve(fs.writeFile(ypath, whoisYaml));
+    }
   });
+}
+
+/**
+ * Takes the string from whois.lookup and parses it into an object.
+ * @param {string} whoisdata
+ * @return {object}
+ */
+function simplifyWhois(whoisdata) {
+  console.info(`${typeof whoisdata} received.`, whoisdata);
+  const whoisObject = whoisdata.split(/\n/);
+  console.info(whoisObject.length, "entries in whoisdata");
+  if (whoisdata.endsWith("was refused.")) {
+    console.error(whoisdata);
+  }
+  if (Object.keys(whoisObject).length <= 10) {
+    return { raw: whoisdata };
+  }
+  const simplifiedObject = {
+    registry: {},
+    registrar: {},
+    registrant: {},
+    registrant_admin: {},
+    registrant_tech: {},
+    name_server: [],
+    domain_status: [],
+    unmatched_data: [],
+  };
+  for (let i = 0; i < whoisObject.length; i++) {
+    const matched = whoisObject[i].match(ATTRIBUTE_REGEX);
+    if (matched && matched.length > 1) {
+      switch (matched[1]) {
+        case "Domain name":
+        case "Domain Name":
+          simplifiedObject.domain_name = matched[2];
+          break;
+        case "Registry Domain ID":
+          simplifiedObject.registry.domain_id = matched[2];
+          break;
+        case "Registry Registrant ID":
+          simplifiedObject.registry.registrant_id = matched[2];
+          break;
+        case "Registry Admin ID":
+          simplifiedObject.registry.admin_id = matched[2];
+          break;
+        case "Registry Tech ID":
+          simplifiedObject.registry.tech_id = matched[2];
+          break;
+        case "Registrar WHOIS Server":
+          simplifiedObject.registrar.whois = matched[2];
+          break;
+        case "Registrar URL":
+          simplifiedObject.registrar.url = matched[2];
+          break;
+        case "Updated Date":
+          simplifiedObject.updated_date = matched[2];
+          break;
+        case "Creation Date":
+          simplifiedObject.created_date = matched[2];
+          break;
+        case "Registrar Registration Expiration Date":
+          simplifiedObject.registrar.registration_expiration = matched[2];
+          break;
+        case "Registrar":
+          simplifiedObject.registrar.name = matched[2];
+          break;
+        case "Registrar Abuse Contact Email":
+          simplifiedObject.registrar.abuse_email = matched[2];
+          break;
+        case "Registrar Abuse Contact Phone":
+          simplifiedObject.registrar.abuse_phone = matched[2];
+          break;
+        case "Registrar IANA ID":
+          simplifiedObject.registrar.iana_id = matched[2];
+          break;
+        case "Domain Status":
+          simplifiedObject.domain_status.push(matched[2]);
+          break;
+        case "Name Server":
+          simplifiedObject.name_server.push(matched[2]);
+          break;
+        case "Reseller":
+          simplifiedObject.reseller = matched[2];
+          break;
+        case "Registrant Name":
+          simplifiedObject.registrant.name = matched[2];
+          break;
+        case "Registrant Organization":
+          simplifiedObject.registrant.organization = matched[2];
+          break;
+        case "Registrant Street":
+          simplifiedObject.registrant.street = matched[2];
+          break;
+        case "Registrant City":
+          simplifiedObject.registrant.city = matched[2];
+          break;
+        case "Registrant State/Province":
+          simplifiedObject.registrant.state_province = matched[2];
+          break;
+        case "Registrant Postal Code":
+          simplifiedObject.registrant.postal_code = matched[2];
+          break;
+        case "Registrant Country":
+          simplifiedObject.registrant.country = matched[2];
+          break;
+        case "Registrant Phone":
+          simplifiedObject.registrant.phone = matched[2];
+          break;
+        case "Registrant Phone Ext":
+          simplifiedObject.registrant.phone_ext = matched[2];
+          break;
+        case "Registrant Fax":
+          simplifiedObject.registrant.fax = matched[2];
+          break;
+        case "Registrant Fax Ext":
+          simplifiedObject.registrant.fax_ext = matched[2];
+          break;
+        case "Registrant Email":
+          simplifiedObject.registrant.email = matched[2];
+          break;
+        case "Admin Name":
+          simplifiedObject.registrant_admin.name = matched[2];
+          break;
+        case "Admin Organization":
+          simplifiedObject.registrant_admin.organization = matched[2];
+          break;
+        case "Admin Street":
+          simplifiedObject.registrant_admin.street = matched[2];
+          break;
+        case "Admin City":
+          simplifiedObject.registrant_admin.city = matched[2];
+          break;
+        case "Admin State/Province":
+          simplifiedObject.registrant_admin.state_province = matched[2];
+          break;
+        case "Admin Postal Code":
+          simplifiedObject.registrant_admin.postal_code = matched[2];
+          break;
+        case "Admin Country":
+          simplifiedObject.registrant_admin.country = matched[2];
+          break;
+        case "Admin Phone":
+          simplifiedObject.registrant_admin.phone = matched[2];
+          break;
+        case "Admin Phone Ext":
+          simplifiedObject.registrant_admin.phone_ext = matched[2];
+          break;
+        case "Admin Fax":
+          simplifiedObject.registrant_admin.fax = matched[2];
+          break;
+        case "Admin Fax Ext":
+          simplifiedObject.registrant_admin.fax_ext = matched[2];
+          break;
+        case "Admin Email":
+          simplifiedObject.registrant_admin.email = matched[2];
+          break;
+        case "Tech Name":
+          simplifiedObject.registrant_tech.name = matched[2];
+          break;
+        case "Tech Organization":
+          simplifiedObject.registrant_tech.organization = matched[2];
+          break;
+        case "Tech Street":
+          simplifiedObject.registrant_tech.street = matched[2];
+          break;
+        case "Tech City":
+          simplifiedObject.registrant_tech.city = matched[2];
+          break;
+        case "Tech State/Province":
+          simplifiedObject.registrant_tech.state_province = matched[2];
+          break;
+        case "Tech Postal Code":
+          simplifiedObject.registrant_tech.postal_code = matched[2];
+          break;
+        case "Tech Country":
+          simplifiedObject.registrant_tech.country = matched[2];
+          break;
+        case "Tech Phone":
+          simplifiedObject.registrant_tech.phone = matched[2];
+          break;
+        case "Tech Phone Ext":
+          simplifiedObject.registrant_tech.phone_ext = matched[2];
+          break;
+        case "Tech Fax":
+          simplifiedObject.registrant_tech.fax = matched[2];
+          break;
+        case "Tech Fax Ext":
+          simplifiedObject.registrant_tech.fax_ext = matched[2];
+          break;
+        case "Tech Email":
+          simplifiedObject.registrant_tech.email = matched[2];
+          break;
+        case "DNSSEC":
+          simplifiedObject.dnssec = matched[2];
+          break;
+        case ">>> Last update of WHOIS database":
+          simplifiedObject.whois_db_update_time = matched[2].match(
+            /^[^< ]+/
+          )[0];
+          break;
+        case "URL of the ICANN WHOIS Data Problem Reporting System":
+          // Do nothing
+          break;
+        default:
+          simplifiedObject.unmatched_data.push({
+            attribute: matched[1],
+            value: matched[2],
+          });
+          break;
+      }
+    } else if (whoisObject[i].startsWith("Access to")) {
+      simplifiedObject.tos = whoisObject[i];
+    }
+  }
+  return simplifiedObject;
 }
