+++ lib/yamler.js
@@ -1,5 +1,4 @@
 const yaml = require("js-yaml");
-const fs = require("fs");
 const path = require("path");
 
 const CONFIG_YAML_PATH = path.join(__dirname, "..", "config", "config.yaml");
@@ -7,43 +6,35 @@ const DOMAIN_YAML_PATH = path.join(__dirname, "..", "config", "domain.yaml");
 const WHOIS_DIR_PATH = path.join(__dirname, "..", "whois-data");
 
 /**
- * Read a yaml file and turn it into a js object
- * @param {string} yamlpath
- * @param {function} callback
+ * Give this some YAML and it resolves to a JS Object
+ * @param {string} yamldoc
+ * @return {Promise}
  */
-function getObjectFromYaml(yamlpath, callback) {
-  try {
-    const doc = yaml.safeLoad(fs.readFileSync(yamlpath, "utf8"));
-    callback(null, doc);
-  } catch (e) {
-    callback(e);
-  }
+function parseObjectFromYaml(yamldoc) {
+  const promisedObject = function (resolve, reject) {
+    try {
+      resolve(yaml.safeLoad(yamldoc));
+    } catch (e) {
+      console.error(`YAML: Unable to parse from doc:`);
+      console.error(yamldoc);
+      reject(e);
+    }
+  };
+  return new Promise(promisedObject);
 }
 
 /**
- * Take a js object and write it into a yaml file
- * @param {string} yamlpath
- * @param {object} object
- * @param {function} callback
+ * Gives YAML string from a js object
+ * @param {object} obj
+ * @return {string}
  */
-function setYamlFromObject(yamlpath, object, callback) {
-  console.log("attempting yaml safe dump before writing");
-  const safeYaml = yaml.safeDump(object);
-  console.log(safeYaml);
-  try {
-    fs.writeFileSync(yamlpath, safeYaml);
-    callback();
-  } catch (e) {
-    console.error(
-      `Caught error writing ${JSON.stringify(object)} to ${yamlpath}`
-    );
-    callback(e);
-  }
+function parseYamlFromObj(obj) {
+  return yaml.safeDump(obj);
 }
 
 module.exports = {
-  getObjectFromYaml,
-  setYamlFromObject,
+  parseObjectFromYaml,
+  parseYamlFromObj,
   CONFIG_YAML_PATH,
   DOMAIN_YAML_PATH,
   WHOIS_DIR_PATH,
