+++ package.json
@@ -19,7 +19,7 @@
     "morgan": "^1.10.0",
     "nodemailer": "^6.4.11",
     "socket.io": "^2.3.0",
-    "whois": "^2.13.0"
+    "whois-promise": "^1.2.4"
   },
   "devDependencies": {
     "eslint": "^7.9.0",
